fix inheritance unlocking for items that don't have groups applied to them -> reason for that is to unlock certain child items

+++ src/MembersBundle/Controller/Admin/RestrictionController.php
@@ -136,8 +136,14 @@ class RestrictionController extends AdminController
 
         //remove restriction since no group is selected any more.
         if (empty($settings->membersDocumentUserGroups)) {
-            if ($hasRestriction === true) {
+            if ($hasRestriction === true && !$restriction->getIsInherited()) {
                 $restriction->delete();
+            }
+            else {
+                $restriction->setInherit($membersDocumentInheritable);
+                $restriction->setIsInherited(false);
+                $restriction->setRelatedGroups($membersDocumentUserGroups);
+                $restriction->save();
             }
             //update or set restriction
         } else {
